requirements.txt
fastmcp==2.13.3
sqlite-vec>=0.1.1
openai>=1.0.0
python-dotenv>=1.0.0
paramiko>=3.4.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
snowflake-connector-python>=3.0.0
